% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.zoocat}
\alias{aggregate.zoocat}
\title{Aggregate \code{zoocat} object}
\usage{
\method{aggregate}{zoocat}(x, by, FUN = mean, ...)
}
\arguments{
\item{x}{a \code{zoocat} object.}

\item{by}{a character string indicates the field of column attributes.}

\item{FUN}{a function to be applied to all data subsets.}

\item{...}{additional arguments to be passed to methods.}
}
\description{
Aggregate columns of \code{zoocat} object by given column attributes fields.
}
\examples{
data(sst)
sstmelt <- melt(sst, id.vars = c('year', 'month'))
zc <- cast2zoocat(sstmelt, index.var = 'year', value.var = 'value')
zc.agg1 <- aggregate(zc, by = 'month')
zc.agg2 <- aggregate(zc, by = 'variable')
}

