% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmonClass.R
\name{as.gmon}
\alias{as.gmon}
\title{Coercion from and to \code{gmon} class}
\usage{
as.gmon(x, ...)
}
\arguments{
\item{x}{an numeric vector.}

\item{...}{unused.}
}
\value{
a "\code{gmon}" object.
}
\description{
Coercion from and to \code{gmon} class
}

