% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.seasonal.R
\name{as.seasonal}
\alias{as.seasonal}
\alias{as.seasonal.mlydata}
\title{Transforming a mlydata Object to be Seasonal Data}
\usage{
as.seasonal(x, DJF.first = TRUE)

\method{as.seasonal}{mlydata}(x, DJF.first = TRUE)
}
\arguments{
\item{x}{A \code{mlydata} object.}

\item{DJF.first}{If TRUE, the winter season will be in the first column.}
}
\value{
A \code{zoo} object.
}
\description{
Transforming a mlydata object to be seasonal data.
}
\examples{
x <- matrix(1 : 36, nrow = 3)
md <- mlydata(x, year = 1991 : 1993, month = 1 : 12)
print(as.seasonal(md))
}

