% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.series.R
\name{as.series}
\alias{as.series}
\alias{as.series.mlydata}
\title{Transforming a mlydata Object into a Series}
\usage{
as.series(x)

\method{as.series}{mlydata}(x)
}
\arguments{
\item{x}{A mlydata object. The month of it must be 1 : 12.}
}
\value{
A series, whose index is a yearmon object definited in the zoo
package.
}
\description{
Transforming a mlydata object into a series of monthly observations.
}
\examples{
x <- matrix(1 : 36, nrow = 3)
md <- mlydata(x, year = 1991 : 1993, month = 1 : 12)
as.series(md)
}

