% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.zoocat.R
\name{as.zoocat}
\alias{as.zoocat}
\alias{as.zoocat.mlydata}
\alias{as.zoocat.mlydataList}
\alias{as.zoocat.zoo}
\title{Coercing Objects to Class \code{zoocat}}
\usage{
as.zoocat(x, ...)

\method{as.zoocat}{mlydata}(..., addname = TRUE, variable.name = "variable")

\method{as.zoocat}{mlydataList}(x, variable.name = "variable", ...)

\method{as.zoocat}{zoo}(x, colattr = NULL, variable.name = "variable", ...)
}
\arguments{
\item{x}{the object.}

\item{...}{other arguments passed to methods.}

\item{addname}{logical. If TRUE, a field of \code{cattr} "name" will be add.}

\item{variable.name}{the name of the field in the \code{cattr} of
the output \code{zoocat} object to store the variable name.
Only valid when \code{addname} is TRUE.}

\item{colattr}{The column attribute table for x.}
}
\value{
A \code{zoocat} object.
}
\description{
Coercing objects to class \code{zoocat}.




}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
md2 <- md + 100
as.zoocat(md)
as.zoocat(md, addname = FALSE)
as.zoocat(x = md)
as.zoocat(x = md, y = md2)
as.zoocat(md, md2)

x <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
y <- x + 100
mdList <- mlydataList(list(x = x, y = y))
as.zoocat(mdList)

zobj <- zoo(matrix(1:10, nrow = 5), order.by = 11:15)
colnames(zobj) <- c('a', 'b')
as.zoocat(zobj)
}

