% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.zoomly.R
\name{as.zoomly}
\alias{as.zoomly}
\alias{as.zoomly.zoo}
\alias{as.zoomly.zoocat}
\title{Coercion Objects to Class \code{zoomly}}
\usage{
as.zoomly(x, ...)

\method{as.zoomly}{zoo}(x, month, ...)

\method{as.zoomly}{zoocat}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{further arguments passed to methods.}

\item{month}{a vector of month, must have same length as ncol of x.}
}
\value{
A zoomly object.
}
\description{
Coercion objects to class \code{zoomly}.


}
\examples{
x <-  matrix(1 : 36, nrow = 3)
zobj <- zoo(x, order.by = 1991 : 1993)
as.zoomly(zobj, month = 1 : 12)

zc <- zoocat(x, order.by = 1991 : 1993,
             colattr = data.frame(month = 1 : 12))
as.zoomly(zc)
}

