% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cast2mlydata.R
\name{cast2mlydata}
\alias{cast2mlydata}
\title{Cast a Data Frame as a \code{mlydata} Object}
\usage{
cast2mlydata(x, year.var = "year", month.var = "month", value.var,
  variable.var = NULL, fun.aggr = NULL)
}
\arguments{
\item{x}{A data frame.}

\item{year.var}{The name of the column to be treated as the index of the mlydata object.}

\item{month.var}{The name of the column which stores month.}

\item{value.var}{The name of the column which stores the values.}

\item{variable.var}{The name of the column which stores the variable names.}

\item{fun.aggr}{Aggregation function needed if variables do not identify a single observation
for each output cell. Defaults to length (with a message) if needed but not specified.}
}
\value{
A \code{mlydata} object.
}
\description{
Cast a data frame as a \code{mlydata} object.
}
\examples{
x <- data.frame(year = rep(1991 : 1993, each = 3), month = rep(3 : 1, 3), value = 1 : 9)
md <- cast2mlydata(x, index.var = 'year', month.var = 'month', value.var = 'value', fun.aggr = mean)

data(sst)
cast2mlydata(sst, value.var = c('nino12', 'nino3'))

sst.melt <- melt(sst, id.var = c('year', 'month'))
sst.cast <- cast2mlydata(sst.melt, value.var = 'value', variable.var = 'variable')
sst.remelt <- melt(sst.cast)
}

