% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cast2mlydata.R
\name{cast2zoomly}
\alias{cast2zoomly}
\alias{cast2zoomly.data.frame}
\alias{cast2zoomly.zoo}
\title{Cast objects to \code{zoomly} or \code{zoomlyList} Objects}
\usage{
cast2zoomly(x, ...)

\method{cast2zoomly}{data.frame}(x, year.var = "year", month.var = "month",
  value.var = NULL, variable.var = NULL, fun.aggregate = NULL, ...)

\method{cast2zoomly}{zoo}(x, value.var = NULL, fun.aggregate = NULL, ...)
}
\arguments{
\item{x}{a object.}

\item{...}{further arguments.}

\item{year.var}{the name of the column which stores year.}

\item{month.var}{the name of the column which stores month.}

\item{value.var}{the name of the column which stores values. Can have
several elements. NULL means all columns except \code{year.var} and \code{month.var}.}

\item{variable.var}{the name of the column which stores the variable names.}

\item{fun.aggregate}{aggregation function needed if variables do not identify a single observation
for each output cell. Defaults to length (with a message) if needed but not specified.
See \code{\link{dcast}}.}
}
\value{
A \code{zoomly} object.
}
\description{
Cast a data frame or \code{zoo} object as a \code{zoomly} or \code{zoomlyList} object.
}
\details{
If the length of \code{value.var} is larger than 1, each variable in \code{value.var}
will be casted as a \code{zoomly} object, and a \code{zoomlyList} will be returned.\cr
If the \code{variable.var} is specified, each value in \code{variable.var} corresponding
to a variable, and corresponding row will be casted into a \code{zoomly} object in
the returned \code{zoomlyList} object.
}
\examples{
x <- data.frame(year = rep(1991 : 1993, each = 3), month = rep(3 : 1, 3),
                value = 1 : 9)
zm <- cast2zoomly(x, year.var = 'year', month.var = 'month',
                   value.var = 'value', fun.aggregate = mean)

data(sst)
cast2zoomly(sst, value.var = c('nino12', 'nino3'))

sst.melt <- melt(sst, id.var = c('year', 'month'))
sst.cast <- cast2zoomly(sst.melt, value.var = 'value', variable.var = 'variable')
sst.remelt <- melt(sst.cast)

ym <- as.yearmon(2000 + seq(0, 23)/12)
zooobj <- zoo(matrix(1:48, nrow = 24), order.by = ym)
colnames(zooobj) <- c('x', 'y')
cast2zoomly(zooobj)
}

