% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cbind.R
\name{cbind}
\alias{cbind}
\alias{cbind.mlydata}
\alias{cbind.zoocat}
\title{Combine \code{zoocat} or \code{mlydata} Objects by Columns}
\usage{
## S3 method for class "zoocat"
cbind(...)

## S3 method for class "mlydata"
cbind(...)
}
\arguments{
\item{...}{\code{zoocat} or \code{mlydata} objects.}
}
\value{
\code{cbind.zoocat} will return a \code{zoocat} object. \code{cbind.mlydata} will return a \code{zoo} object.
}
\description{
Combine \code{zoocat} or \code{mlydata} objects by columns.
}
\details{
cbind.zoocat is similar with cbind.zoo, and the column attributes are also
combined. For the combination of cattr, some NA will be filled in if it is
necessary. cbind.mlydata will return a \code{object}.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = 'sst')
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
x2 <- x + 100
colAttr2 <- data.frame(level = 500, month = c(4, 6, 7, 9))
zc2 <- zoocat(x2, order.by = 1991 : 1995, colattr = colAttr2)
zc3 <- cbind(zc, zc2)
cattr(zc3)

x <- matrix(1 : 20, nrow = 5)
md1 <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
md2 <- md1 + 1
cbind(md1, md2)
}

