% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor}
\alias{cor}
\alias{cor.zoo}
\alias{cor.zoocat}
\title{Correlation Computing for \code{zoo} or \code{zoocat} Objects}
\usage{
cor(x, y = NULL, ...)

\method{cor}{zoo}(x, y = NULL, ...)

\method{cor}{zoocat}(x, y = NULL, ...)
}
\arguments{
\item{x}{A \code{zoo} or \code{zoocat} object.}

\item{y}{NULL or a \code{zoo} or \code{zoocat} object. If NULL, x will be used.}

\item{...}{Other arguments for function \code{cor}.}
}
\description{
Compute corelations for \code{zoo} or \code{zoocat} objects.
}
\examples{

x <- zoo(c(1, 3, 2, 4, 5))
y <- zoo(c(12, 30, 2, 46, 5))
cor(x, y)

x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = 'x')
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
y <- zoo(c(3, 5, 4, 6, 3), order.by = 1991 : 1995)
cor(zc, y, method = 'kendall')
cor(zc, method = 'kendall')


}

