% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_apply.R
\name{core_apply}
\alias{core_apply}
\alias{core_apply.zoocat}
\title{Apply Functions Over the Core Matrix}
\usage{
core_apply(x, ...)

\method{core_apply}{zoocat}(x, fun, bind, ...)
}
\arguments{
\item{x}{the object.}

\item{...}{other arguments for \code{fun}.}

\item{fun}{the function to apply. The \code{fun} must return a matrix or
a vector.}

\item{bind}{a vector of length 1 or 2 with element values to be
 'cattr' or 'index' or NA to describe how to bind the return
data with \code{cattr} or \code{index}. If \code{fun} return a vector, set
\code{bind} to be a scalar. If \code{fun} return a matrix, set
\code{bind} to be a vector of length 2. See details.}
}
\description{
Apply a function over the core matrix of the \code{zoocat} object,
and bind the return data with \code{cattr} or \code{index}.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c(rep('xxx', 3), 'yyy'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
core_apply(zc, fun = colMeans, bind = 'cattr')
core_apply(zc, fun = rowMeans, bind = 'index')
core_apply(zc, fun = function (x) {x*2}, bind = c('index', 'cattr'))
core_apply(zc, fun = function (x) {t(x*2)}, bind = c('cattr', 'index'))
core_apply(zc, fun = function (x) {x*2}, bind = c('index', NA))
core_apply(zc, fun = function (x) {x[3:4, ]}, bind = c(NA, 'cattr'))
core_apply(zc, fun = function (x)
                      {r <- x[3:4, ]
                      rownames(r) <- c('a', 'b')
                      return(r)},
                      bind = c(NA, 'cattr'))

vec <- as.vector(zc[, 1])
core_apply(zc, fun = function (x) {cor(x, vec)}, bind = 'cattr')
}

