% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\alias{corr.zoo}
\alias{corr.zoocat}
\title{Correlation Computing for \code{zoo} or \code{zoocat} Objects}
\usage{
corr(x, y, ...)

\method{corr}{zoo}(x, y, ...)

\method{corr}{zoocat}(x, y, ...)
}
\arguments{
\item{x}{A \code{zoo} or \code{zoocat} object.}

\item{...}{Other arguments for function \code{cor}.}

\item{y}{A \code{zoo} object with the same index as \code{x}.}
}
\value{
For \code{corr.zoo}, the same as in function \code{cor}. For
\code{corr.zoocat}, a data frame will be returned.
}
\description{
Compute correlations for \code{zoo} or \code{zoocat} objects.
}
\examples{
x <- zoo(c(1, 3, 2, 4, 5))
y <- zoo(c(12, 30, 2, 46, 5))
corr(x, y)

x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = 'sst')
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
y <- zoo(c(3, 5, 4, 6, 3), order.by = 1991 : 1995)
corr(zc, y)
}

