% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expand.R, R/methods_mlydataList.R
\name{expand}
\alias{cummax.mlydataList}
\alias{expand}
\alias{expand.mlydata}
\title{Expand \code{mlydata} objects}
\usage{
expand(x, ...)

\method{expand}{mlydata}(x, k, direction = "both", ...)

\method{cummax}{mlydataList}(x, ...)
}
\arguments{
\item{x}{the \code{mlydata} object.}

\item{...}{further arguments.}

\item{k}{the number of cycle to be expanded.}

\item{direction}{the direction to expand. Can be "left",
"right" or "both".}
}
\value{
a \code{mlydata} object.
}
\description{
Expand \code{mlydata} objects
}
\examples{
x <- matrix(1 : 48, nrow = 4)
md <- mlydata(x, year = 1991 : 1994)
expand(md, k = 1, direction = 'left')
expand(md, k = 1, direction = 'right')
expand(md, k = 1, direction = 'both')
expand(md, k = 2, direction = 'both')
}

