% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_zoocat.R
\name{gmon_trans}
\alias{gmon_trans}
\alias{scale_x_gmon}
\alias{scale_y_gmon}
\title{Tools for plottin \code{zoocat} objects with ggplot2}
\usage{
gmon_trans(n = 5)

scale_x_gmon(..., n = 5)

scale_y_gmon(..., n = 5)
}
\description{
Tools for plottin \code{zoocat} objects with ggplot2
}
\examples{
library(ggplot2)
df <- data.frame(month = gmon(-5:5), value = 1:11)
ggplot(df, aes(month, value)) + 
    geom_line() + geom_point() +
    scale_x_gmon()
    
}

