% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{grading}
\alias{grading}
\alias{grading.default}
\alias{grading.zoo}
\title{Grading Series}
\usage{
grading(x, ...)

\method{grading}{default}(x, alloc, boundary.ret = FALSE, grade.name = NULL,
  ...)

\method{grading}{zoo}(x, ...)
}
\arguments{
\item{x}{a vector or a zoo object.}

\item{...}{further arguments.}

\item{alloc}{the numbers of elements in each grade. sum(alloc) must ==
length(x).}

\item{boundary.ret}{logical. If TRUE, the boundaries between adjacent grades will be
returned.}

\item{grade.name}{character string. If NULL, the names of grades will be set to be 1, 2, ... ,
N. The 1st element of grade.name represents the grade of smallest value.}
}
\value{
If boundary.ret = TRUE, a list will be returned. The first element
is the grades of each element in x. The second is the boundaries between
grades. If boundary.ret = FALSE, only grades will be returned.
}
\description{
Grading series by order of each element.
}
\examples{

x <- zoo(1:10)
grading(x, alloc = c(2, 2, 6))
grading(x, alloc = c(2, 2, 6), grade.name = c('a', 'b', 'c'))
grading(x, alloc = c(2, 2, 6), grade.name = c(3, 3, 1))
grading(x, alloc = c(2, 2, 6), boundary.ret = TRUE)

x <- 1 : 10
grading(x, alloc = c(2, 2, 6), grade.name = c('a', 'b', 'c'))

}

