% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/index.name.R
\name{index.name}
\alias{index.name}
\alias{index.name.zoocat}
\alias{index.name<-}
\alias{index.name<-.zoocat}
\title{Get or set the name of the index variable}
\usage{
index.name(x)

index.name(x) <- value

\method{index}{name.zoocat}(x)

\method{index}{name<-.zoocat}(x, value)
}
\arguments{
\item{x}{the object.}

\item{value}{the new value.}
}
\description{
Get or set the name of the index variable
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c(rep('xxx', 3), 'yyy'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
index.name(zc)
index.name(zc) <- 'time'
}

