% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indname.R
\name{indname}
\alias{indname}
\alias{indname.zoocat}
\alias{indname<-}
\alias{indname<-.zoocat}
\title{Get or set the name of the index variable}
\usage{
indname(x)

indname(x) <- value

\method{indname}{zoocat}(x)

\method{indname}{zoocat}(x) <- value
}
\arguments{
\item{x}{the object.}

\item{value}{the new value.}
}
\description{
Get or set the name of the index variable
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c(rep('xxx', 3), 'yyy'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
indname(zc)
indname(zc) <- 'time'

}

