% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isvalid.R
\name{isvalid}
\alias{isvalid}
\alias{isvalid.zoocat}
\title{Determine the Validity of a \code{zoocat} Object}
\usage{
isvalid(x)

\method{isvalid}{zoocat}(x)
}
\arguments{
\item{x}{A \code{zoocat} object.}
}
\value{
A logic variable.
}
\description{
Determine the validity of a \code{zoocat} object.
}
\examples{

x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = 'sst')
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
isvalid(zc)

}

