% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lag.R
\name{lag.zoomly}
\alias{lag.zoomly}
\title{Lags of \code{zoomly} Objects}
\usage{
\method{lag}{zoomly}(x, k = 1, adjust.month = TRUE, ...)
}
\arguments{
\item{x}{A \code{zoocat} or \code{zoomly} object.}

\item{k}{The number of the lags.}

\item{adjust.month}{Logical. If TRUE, the month of \code{x} will be
add \code{12 * k}.}

\item{...}{further arguments.}
}
\description{
Lags of \code{zoomly} Objects
}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- zoomly(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
lag(md, k = 1)
lag(md, k = -2)
lag(md, k = -1)
cbind(lag(md, -1), md)

lag(md, k = -1, adjust.month = FALSE)
}

