% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/margin.R
\name{margin}
\alias{margin}
\alias{margin.data.frame}
\alias{margin.zoocat}
\title{Computing Margins}
\usage{
margin(x, ...)

\method{margin}{data.frame}(x, mar.for, value.var = "value",
  fun.aggr = mean)

\method{margin}{zoocat}(x, mar.for, fun.aggr = mean)
}
\arguments{
\item{x}{A data frame or \code{zoocat} object.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{mar.for}{One or more variables to compute margin for.
For data frame, it means column names. For \code{zoocat} object,
it means names of column attributes.}

\item{value.var}{The value variable.}

\item{fun.aggr}{The function for aggregation.}
}
\description{
Computing margins for given ID variables of a data frame
 or given column attributes of a \code{zoocat} object.




}
\examples{
## The example of data frame
lat <- 1:5
lon <- 101:105
time <- 1991:2000
data <- data.frame(lat = rep(lat, 50), lon = rep(rep(lon, each = 5), 10),
                   time = rep(time, each = 25), value = 1 : 250)
agg1 <- margin(data, mar.for = c('lat'))
agg2 <- margin(data, mar.for = c('lon', 'lat'))
agg3 <- margin(data, mar.for = c('lon', 'lat'), fun.aggr = sum)
agg4 <- margin(data, mar.for = c('lon'), fun.aggr = sum)

## The example of zoocat
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c('a', 'b'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
zc.mar <- margin(zc, mar.for = 'month')
}

