% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/melt.zoocat.R
\name{melt.zoocat}
\alias{melt.zoocat}
\title{Melt a \code{zoocat} Object}
\usage{
\method{melt}{zoocat}(x, value.name = "value", index.name = "index")
}
\arguments{
\item{x}{A \code{zoocat} object.}

\item{value.name}{Name of variable used to store values.}

\item{index.name}{Name of variable used to store the index of the \code{zoocat} object.}
}
\value{
A data frame.
}
\description{
Melt a \code{zoocat} object to a data frame of the long table style,
which is similar as in package \code{reshape2}.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c('x', 'y', 'z', 'a'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
melt(zc)
}

