% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlydata.R
\name{mlydata}
\alias{mlydata}
\alias{mon}
\alias{mon.mlydata}
\alias{mon<-}
\alias{mon<-.mlydata}
\alias{yr}
\alias{yr.mlydata}
\alias{yr<-}
\alias{yr<-.mlydata}
\title{\code{mlydata} Class}
\usage{
mlydata(x, year, month = 1:12, order.by = NULL)

mon(x)

mon(x) <- value

yr(x)

yr(x) <- value

\method{mon}{mlydata}(x)

\method{yr}{mlydata}(x)

\method{mon}{mlydata}(x) <- value

\method{yr}{mlydata}(x) <- value
}
\arguments{
\item{x}{For \code{mlydata} function, it is a matrix or a vector. If x is
a matrix, each row will be treated as a year. If x is a vector, it will be
treated as a matrix with only one column. For other methods, x is a
\code{mlydata} object.}

\item{year}{A numeric vector representing years.}

\item{month}{A numeric vector in which numbers are used to represent months.
The elements of \code{month} can be smaller than 1 or larger than 12, which
means the months of past years or future years. For example, the months of
last year is (1 : 12) - 12.}

\item{order.by}{the same as year.}

\item{value}{The new value for month and year of the object.}
}
\value{
\code{mlydata} returns a \code{mlydata} object.
}
\description{
A class designed for monthly data.
}
\details{
\code{mlydata} class is dependent on \code{zoo} class. An attribute
"month" is added for storing the month information.\cr For \code{print}
method, each column name is set as month abbreviation + month number. For
example, the column name corresponding to Feb of current year is "Feb.2".
When all month numbers are between 1 and 12, only the abbreviations of months is
printed.\cr Using \code{yr} and \code{mon} to get and set the year and the
month of the object.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
coredata(md)
mon(md)
mon(md) <- mon(md) + 1
yr(md)
yr(md) <- yr(md) + 1
md[3, 4]
md[3, , drop = FALSE]
md[, 2]
md['1994', ]
md[, 'Apr']

x <-  matrix(1 : 36, nrow = 3)
md <- mlydata(x, year = 1991 : 1993)
}

