% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlydataList.R
\name{mlydataList}
\alias{mlydataList}
\title{mlydataList Class}
\usage{
mlydataList(x)
}
\arguments{
\item{x}{A list, each elements of which is a mlydata object. x must has
names for each elements.}
}
\value{
A mlydataList object.
}
\description{
mlydataList class.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
md1 <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
md2 <- md1 + 1
mdList <- mlydataList(list(md1 = md1, md2 = md2))
as.zoocat(mdList)
}

