% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order_col.R
\name{order_col}
\alias{order_col}
\title{Order by column}
\usage{
order_col(x)
}
\description{
Order by column
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c(rep('xxx', 3), 'yyy'))
zm <- zoomly(x, order.by = 1991 : 1995, colattr = colAttr)
order_col(zm)
zm <- zm[, c(4, 2, 3, 1)]
order_col(zm)
}

