% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmonClass.R
\name{rela_year}
\alias{rela_year}
\title{Get the relative years for a \code{gmon} object}
\usage{
rela_year(x)
}
\arguments{
\item{x}{a \code{gmon} object.}
}
\description{
Get the relative years for a \code{gmon} object
}
\examples{
num.mon <- -12 : 3
gm <- gmon(num.mon)
ry <- rela_year(gm)
df <- data.frame(num.mon, ry, as.character(gm))
print(df)
}

