% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmeans.zoo.R
\name{rmeans}
\alias{rmeans}
\alias{rmeans.zoo}
\title{Row Means of zoo Object}
\usage{
rmeans(x)

\method{rmeans}{zoo}(x)
}
\arguments{
\item{x}{A zoo object.}
}
\value{
A zoo object.
}
\description{
Row Means of zoo Object
}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
rmeans(md)
}

