% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollmean.mlydata.R
\name{rollmean.mlydata}
\alias{rollmean.mlydata}
\title{The Rolling Means of a mlydata Object}
\usage{
rollmean(x, k, onlyUsePrev = TRUE)
}
\arguments{
\item{x}{A \code{mlydata} object.}

\item{k}{The width of the rolling window.}

\item{onlyUsePrev}{If TRUE, the rolling window is asymmetric, and only
previous and current data is in the rolling window.}
}
\value{
A mlydata object.\cr
}
\description{
The rolling means of a mlydata object.
}
\examples{
x <- matrix(1 : 60, nrow = 5)
md <- mlydata(x, year = 1991 : 1995)
rollmean(md, k = 2)
rollmean(md, k = 3, onlyUsePrev = FALSE)
}

