% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollmean_by_month.mlydata.R
\name{rollmean_by_month}
\alias{rollmean_by_month}
\alias{rollmean_by_month.mlydata}
\title{The Rolling Means of a mlydata Object by Month}
\usage{
rollmean_by_month(x, ...)

\method{rollmean_by_month}{mlydata}(x, k, align = "center")
}
\arguments{
\item{x}{A \code{mlydata} object.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{k}{The width of the rolling window.}

\item{align}{Character specifying whether the index of the result
should be left- or right-aligned or centered (default) compared to
the rolling window of observations.}
}
\value{
A mlydata object.\cr
}
\description{
The rolling means of a mlydata object by month.
}
\examples{
x <- matrix(1 : 60, nrow = 5, byrow = T)
md <- mlydata(x, year = 1991 : 1995)
rollmean_by_month(md, k = 3, align = 'left')
rollmean_by_month(md, k = 3, align = 'right')
rollmean_by_month(md, k = 3)
rollmean(md, k = 3)
}

