% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/row_means.zoo.R
\name{row_means}
\alias{row_means}
\alias{row_means.zoo}
\title{Row Means of zoo Object}
\usage{
row_means(x, ...)

\method{row_means}{zoo}(x, ...)
}
\arguments{
\item{x}{A zoo object.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A zoo object.
}
\description{
Row Means of zoo Object
}
\examples{
x <- matrix(1 : 20, nrow = 5)
zm <- zoomly(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
row_means(zm)
}

