% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rowbind.R
\name{rowbind}
\alias{rowbind}
\title{Combine Data Frames by Rows}
\usage{
rowbind(dfList)
}
\arguments{
\item{dfList}{A list of data frames.}
}
\value{
A data frame.
}
\description{
Combine data frames by rows. The data frames can have different number of
columns or different column names, which is different with \code{rbind}
}
\details{
The columns of data frame should not be class "factor".
}
\examples{
colAttr1 <- data.frame(month = c(2, 3, 5, 6), name = 'sst', stringsAsFactors = F)
colAttr2 <- data.frame(month = c(4, 6, 7, 9), level = 500)
colAttr3 <- data.frame(id = 10 : 14, name = 'slp', stringsAsFactors = F)
rowbind(list(colAttr1, colAttr2, colAttr3))
}

