% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spdf2mat.R
\name{spdf2mat}
\alias{spdf2mat}
\title{Translate Data Frame of Spatial Data to Matrix}
\usage{
spdf2mat(x, lat.var = "lat", lon.var = "lon", value.var = "value")
}
\arguments{
\item{x}{A data frame.}

\item{lat.var}{Name of column which store latitude.}

\item{lon.var}{Name of column which store longitude.}

\item{value.var}{Name of column which store values.}
}
\value{
A list contain matrix and bbox.
}
\description{
Translate a data frame of spatial data to a matrix. The data frame must contain
columns of latitude, longitude and values.
}
\examples{
x <- querydb(fileName = 'ncepncar/hgt500.db',nfetch = -1,
            statement = 'select * from main where month = 6 and year = 2010',
            mar.for = c('lat', 'lon'))
mat <- spdf2mat(x)
}

