% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmonClass.R
\name{true_month}
\alias{true_month}
\title{Get the true month of a \code{gmon} object.}
\usage{
true_month(x)
}
\arguments{
\item{x}{a \code{gmon} object.}
}
\value{
a vector.
}
\description{
Get the true month of a \code{gmon} object.
}
\examples{
num.mon <- -25 : 15
gm <- gmon(num.mon)
tm <- true_month(gm)
cbind(gm, tm)
}

