% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/window.mlydata.R
\name{window.mlydata}
\alias{window.mlydata}
\title{Get Subset of a mlydata Object}
\usage{
window(x, year = NULL, month = NULL)
}
\arguments{
\item{x}{A \code{mlydata} object.}

\item{year}{A numeric vector representing years.}

\item{month}{A numeric vector in which numbers are used to represent months.
The elements of \code{month} can be smaller than 1 or larger than 12, which
means the months of past years or future years. For example, the months of
last year is (1 : 12) - 12.}
}
\value{
A mlydata object.
}
\description{
Get a subset of a mlydata object.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
window(md, year = 1992 : 1993, month = c(3, 6))
}

