% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/window.zoomly.R, R/z_mlydataList_methods.R
\name{window}
\alias{window}
\alias{window.mlydataList}
\alias{window.zoomly}
\title{Get Subset of a \code{zoomly} Object}
\usage{
\method{window}{zoomly}(x, year = NULL, month = NULL, ...)

\method{window}{mlydataList}(x, ...)
}
\arguments{
\item{x}{A \code{zoomly} object.}

\item{year}{A numeric vector representing years.}

\item{month}{A numeric vector in which numbers are used to represent months.
The elements of \code{month} can be smaller than 1 or larger than 12, which
means the months of past years or future years. For example, the months of
last year is (1 : 12) - 12.}

\item{...}{Other arguments for \code{window.zoo}.}
}
\value{
A \code{zoomly} object.
}
\description{
Get a subset of a \code{zoomly} object.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- zoomly(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
window(md, year = 1992 : 1993, month = c(3, 6))
window(md, start = 1993)


data(sst)
mdList <- cast2zoomly(sst, value.var = c('nino12', 'nino3'))
window(mdList, year = 1992:1993, month = 3:5)
}

