% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_col.R
\name{aggregate_col}
\alias{aggregate_col}
\title{Aggregate \code{zoocat} object by column attributes}
\usage{
aggregate_col(x, by, FUN = mean, ...)
}
\arguments{
\item{x}{a \code{zoocat} object.}

\item{by}{a character string indicates the field of column attributes.}

\item{FUN}{a function to be applied to all data subsets.}

\item{...}{additional arguments to be passed to methods.}
}
\description{
Aggregate columns of \code{zoocat} object by given column attributes fields.
}
\examples{
data(nino)
nino.melt <- melt(nino, id.vars = c('year', 'month'))
zc <- cast2zoocat(nino.melt, index.var = 'year', value.var = 'value') 
zc.agg1 <- aggregate_col(zc, by = 'month')
zc.agg2 <- aggregate_col(zc, by = 'variable')

}

