% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apply_col.R
\name{apply_col}
\alias{apply_col}
\alias{apply_col.zoocat}
\title{Apply Functions Over Each Column}
\usage{
apply_col(x, ...)

\method{apply_col}{zoocat}(x, fun, col.as = "vector", ...)
}
\arguments{
\item{x}{A object.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{fun}{The function apply for each column.}

\item{col.as}{If vector, each column will be treated as a vector. If
zoo, each column will be treated as a zoo object.}
}
\description{
Apply a function over each column of the \code{zoocat} object
and return a data frame.
}
\examples{
library(lattice)

data(nino)
ninomelt <- melt(nino, id.var = c('year', 'month'))
zc <- cast2zoocat(ninomelt, index.var = 'year', value.var = 'value')
apply_col(zc, fun = function(x) {mean(x, na.rm = TRUE)}, col.as = 'vector')

fuse <- function (x) {
    mean(window(x, start = 1995, end = 2000))
}
retdf <- apply_col(zc, fun = fuse, col.as = 'zoo')
xyplot(output~month|variable, data = retdf, type = 'o')


fuse <- function (x) {
    x <- window(x, start = 1980, end = 2014)
    ret <- c(max(x), min(x), mean(x), median(x), sd(x))
    names(ret) <- c('max', 'min', 'mean', 'median', 'sd')
    return(ret)
}
retdf <- apply_col(zc, fun = fuse, col.as = 'zoo')
xyplot(sd~month|variable, data = retdf, type = 'o')
}

