% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.mlydata.R
\name{as.mlydata}
\alias{as.mlydata}
\alias{as.mlydata.zoo}
\title{Coercion Objects to Class \code{mlydata}}
\usage{
as.mlydata(x, ...)

\method{as.mlydata}{zoo}(x, month = 1:12)
}
\arguments{
\item{x}{A zoo object.}

\item{month}{A vector of month, must have same length as ncol of x.}
}
\value{
A mlydata object.
}
\description{
Coercion objects to class \code{mlydata}.


}
\examples{
x <-  matrix(1 : 36, nrow = 3)
zobj <- zoo(x, order.by = 1991 : 1993)
md <- as.mlydata(zobj, month = 1:12)

x <-  matrix(1 : 36, nrow = 3)
zcat <- zoocat(x, order.by = 1991 : 1993, colattr = data.frame(month = 1 : 12))
md <- as.mlydata(zcat)
}

