% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.seasonal.R
\name{as.seasonal}
\alias{as.seasonal}
\alias{as.seasonal.zoomly}
\title{Transforming a \code{zoomly} Object to be Seasonal Data}
\usage{
as.seasonal(x, ...)

\method{as.seasonal}{zoomly}(x, DJF.first = TRUE, ...)
}
\arguments{
\item{x}{A \code{zoomly} object.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{DJF.first}{If TRUE, the winter season will be in the first column.}
}
\value{
A \code{zoo} object.
}
\description{
Transforming a \code{zoomly} object to be seasonal data.
}
\examples{
x <- matrix(1 : 36, nrow = 3, byrow = TRUE)
md <- zoomly(x, year = 1991 : 1993, month = 1 : 12)
as.seasonal(md)
as.seasonal(md, DJF.first = FALSE)
}

