% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.zoo.R
\name{as.zoo}
\alias{as.zoo}
\alias{as.zoo.mlydata}
\alias{as.zoo.zoocat}
\alias{as.zooreg.mlydata}
\alias{as.zooreg.zoocat}
\title{Coercion Objects to Class \code{zoo}}
\usage{
\method{as.zoo}{mlydata}(x)

\method{as.zooreg}{mlydata}(x)

\method{as.zoo}{zoocat}(x)

\method{as.zooreg}{zoocat}(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A zoo object.
}
\description{
Coercing objects to class {zoo}.
}
\details{
For methods as.zoo.mlydata and as.zoo.zoocat, the returned zoo object will
be added column names automatically. \cr
Note that the result of \code{as.zoo} will be a \code{zooreg} object if
the input is inherited from \code{zooreg}.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
z <- as.zoo(md)

x <- matrix(1 : 20, nrow = 5)
md1 <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
md2 <- md1 + 1
mdList <- mlydataList(list(md1 = md1, md2 = md2))
zc <- as.zoocat(mdList)
as.zoo(zc)
}

