% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.zoocat.R
\name{as.zoocat}
\alias{as.zoocat}
\alias{as.zoocat.mlydata}
\alias{as.zoocat.mlydataList}
\alias{as.zoocat.zoo}
\title{Coercing Objects to Class \code{zoocat}}
\usage{
as.zoocat(x, ...)

\method{as.zoocat}{mlydata}(x, addname = TRUE, varname = NULL)

\method{as.zoocat}{mlydataList}(x)

\method{as.zoocat}{zoo}(x, colattr = NULL)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{addname}{Logical. If TRUE, a field of \code{cattr} "name" will be add.}

\item{varname}{The value for the name field in the \code{cattr} of
the output \code{zoocat} object. Only valid when \code{addname} is TRUE.
If NULL, the variable name will be used.}

\item{colattr}{The column attribute table for x.}
}
\value{
A \code{zoocat} object.
}
\description{
Coercing objects to class \code{zoocat}.




}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
as.zoocat(md)
as.zoocat(md, addname = F)
as.zoocat(md, varname = 'x')

mat <- matrix(1 : 20, nrow = 5)
x <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
y <- x + 100
mdList <- mlydataList(list(x = x, y = y))
as.zoocat(mdList)

zobj <- zoo(matrix(1:10, nrow = 5), order.by = 11:15)
colnames(zobj) <- c('a', 'b')
as.zoocat(zobj)
}

