% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.zoocat.R
\name{as.zoocat}
\alias{as.zoocat}
\alias{as.zoocat.mlydata}
\alias{as.zoocat.mlydataList}
\title{Coercing Objects to Class "zoocat"}
\usage{
as.zoocat(x, ...)

\method{as.zoocat}{mlydata}(x, varname = NULL)

\method{as.zoocat}{mlydataList}(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A zoocat object.
}
\description{
Coercing objects to class "zoocat".




}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
zc <- as.zoocat(md)
x <- matrix(1 : 20, nrow = 5)
md1 <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
md2 <- md1 + 1
mdList <- mlydataList(list(md1 = md1, md2 = md2))
zc <- as.zoocat(mdList)
}

