% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cast2zoocat.R
\name{cast2zoocat}
\alias{cast2zoocat}
\title{Cast a Data Frame as a zoocat Object}
\usage{
cast2zoocat(x, index.var, value.var, attr.var = NULL, fun.aggr = NULL,
  del.unique.cattr = TRUE)
}
\arguments{
\item{x}{A data frame.}

\item{index.var}{The name of the column to be treated as the index of the
zoocat object.}

\item{value.var}{The name of the column which stored the values.}

\item{attr.var}{The name of the column which will be used as column attributes of
the \code{zoocat} object. If NULL, all columns except \code{value.var} and \code{index.var}
will be used.}

\item{fun.aggr}{Aggregation function needed if variables do not identify a single observation
for each output cell. Defaults to length (with a message) if needed but not specified.}

\item{del.unique.cattr}{If TRUE, the column attibutes with unique value will
be deleted.}
}
\value{
A zoocat object.
}
\description{
Cast a data frame into a zoocat object. The data frame should be the style
of the molten data frame in package reshape2.
}
\examples{
df <- data.frame(year = rep(1991 : 1995, each = 24), month = rep(1 : 12, 10),
                 varname = rep(c('a', 'b'), each = 12), city = rep(1 : 3, each = 40), value = 1 : 120)
cast2zoocat(df, index.var = 'year', value.var = 'value')
cast2zoocat(df, index.var = 'year', value.var = 'value', attr.var = 'varname')

## This is the air quality example from package reshape2
names(airquality) <- tolower(names(airquality))
aqm <- melt(airquality, id=c("month", "day"), na.rm=TRUE)
head(aqm)
cast2zoocat(aqm, index.var = 'month', value.var = 'value', attr.var = 'variable')
cast2zoocat(aqm, index.var = 'month', value.var = 'value')
}

