% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cattr.R
\name{cattr}
\alias{cattr}
\alias{cattr<-}
\title{Getting or Setting the Column Attributes of a zoocat Object.}
\usage{
cattr(x)

cattr(x)
}
\description{
Getting or setting the column attributes of a zoocat object.
}
\details{
Use attr(x, 'cattr') to get or set cattr more flexibly.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = 'sst')
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
colatt <- cattr(zc)
colatt[, 1] <- colatt[, 1] + 1
cattr(zc) <- colatt
}

