% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cattr2name.R
\name{cattr2name}
\alias{cattr2name}
\title{Transforming Data Frame of Column Attributes to Column Names Vector.}
\usage{
cattr2name(colattr)
}
\arguments{
\item{colattr}{A data frame of column attributes in zoocat object.}
}
\value{
A character vector.
}
\description{
Transforming data frame of column attributes to column names vector, which
is useful for zoocat class.
}
\examples{
colAttr <- data.frame(month = c(-2, -3, 5, 6), lat = c(-20, -10, 0, 10))
cattr2name(colAttr)
}

