% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expand.R, R/z_mlydataList_methods.R, R/z_zoomlyList_methods.R
\name{expand}
\alias{expand}
\alias{expand.mlydataList}
\alias{expand.zoomly}
\alias{expand.zoomlyList}
\title{Expand \code{zoomly} objects}
\usage{
expand(x, ...)

\method{expand}{zoomly}(x, left = 1, right = 12, naTrim = TRUE, ...)

\method{expand}{mlydataList}(x, ...)

\method{expand}{zoomlyList}(x, ...)
}
\arguments{
\item{x}{the \code{zoomly} object.}

\item{...}{further arguments input to \code{na.trim}.}

\item{left}{the month of the left limit.}

\item{right}{the month of the right limit.}

\item{naTrim}{logical. If TRUE, na.trim will be used to trim the
result.}
}
\value{
a \code{zoomly} or \code{zoomlyList} object.
}
\description{
Expand \code{zoomly} objects
}
\examples{
x <- matrix(1 : 48, nrow = 4, byrow = TRUE)
zm <- zoomly(x, year = 1991 : 1994)
expand(zm, left = -2)
expand(zm, right = 15)
expand(zm, right = 15, naTrim = FALSE)
expand(zm, left = -13, right = 14)
expand(zm, left = -25, right = 14, naTrim = FALSE)
}

