% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_col.R
\name{filter_col}
\alias{filter_col}
\alias{filter_col_q}
\alias{filter_col_q.zoocat}
\title{Return columns with matching conditions}
\usage{
filter_col_q(x, ...)

\method{filter_col_q}{zoocat}(x, cond)
}
\arguments{
\item{x}{the object.}

\item{...}{other arguments.}

\item{cond}{logical predicates of conditions. Multiple conditions are
combined with &.}
}
\description{
Return columns with matching conditions for the column attributes
table.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c(rep('xxx', 3), 'yyy'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
filter_col_q(zc, quote(month > 2))
temp <- 2
filter_col_q(zc, quote(month == temp))
filter_col_q(zc, quote(month > temp & name == 'yyy'))
}

