% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmonClass.R
\name{gmon}
\alias{gmon}
\title{A class for generalized month index}
\usage{
gmon(x)
}
\description{
A class for generalized month index
}
\examples{
mvec <- gmon(-20:25)
print(mvec)
x <- 1 : 46
names(x) <- mvec
print(x)
}

