% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/margin.R
\name{margin}
\alias{margin}
\alias{margin.data.frame}
\alias{margin.zoocat}
\title{Computing Margins}
\usage{
margin(x, ...)

\method{margin}{data.frame}(x, mar.for, value.var = "value",
  fun.aggr = mean, ...)

\method{margin}{zoocat}(x, mar.for, fun.aggr = mean, ...)
}
\arguments{
\item{x}{A data frame or \code{zoocat} object.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{mar.for}{One or more variables to compute margin for.
For data frame, it means column names. For \code{zoocat} object,
it means names of column attributes.}

\item{value.var}{The value variable.}

\item{fun.aggr}{The function for aggregation.}
}
\description{
Computing margins for given ID variables of a data frame
 or given column attributes of a \code{zoocat} object.




}
\examples{
## The example of data frame
data(sst)
margin(sst, mar.for = c('year'), value.var = 'nino12')
margin(sst, mar.for = c('year'), value.var = 'nino3', fun.aggr = length)

library(lattice)
sstmon <- margin(sst, mar.for = c('month'), value.var = 'nino3')
xyplot(nino3 ~ month, data = sstmon, type = 'o')
sstyr <- margin(sst, mar.for = c('year'), value.var = 'nino3')
xyplot(nino3 ~ year, data = sstyr, type = 'o')

sstmelt <- melt(sst, id.var = c('year', 'month'))
sstMonVar <- margin(sstmelt, mar.for = c('month', 'variable'), value.var = 'value')
xyplot(value~month, data = sstMonVar, groups = variable, type = 'o')

## The example of zoocat
zc <- cast2zoocat(sstmelt, index.var = 'year', value.var = 'value')
zcmar1 <- margin(zc, mar.for = 'month')
zcmar2 <- margin(zc, mar.for = 'variable')
plot(zcmar2)
}

