% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/melt.R
\name{melt}
\alias{melt}
\alias{melt.mlydata}
\alias{melt.zoocat}
\title{Melt a \code{zoocat} or \code{mlydata} Object}
\usage{
\method{melt}{zoocat}(x, value.name = "value", index.name = "index",
  na.rm = FALSE, ...)

\method{melt}{mlydata}(x, value.name = "value", ret = "data.frame",
  na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{zoocat} object.}

\item{value.name}{Name of variable used to store values.}

\item{index.name}{Name of variable used to store the index of the \code{zoocat} object.}

\item{na.rm}{As \code{melt} in reshape2.Should NA values be removed from the data set?
For \code{mlydata}, it is only valid when \code{ret} is \code{data.frame}.}

\item{...}{Further arguments.}

\item{ret}{Can be \code{data.frame} or \code{zoo}}
}
\value{
A data frame or a \code{zoo} object.
}
\description{
Melt a \code{zoocat} object or \code{mlydata} to a data frame of the long table style,
which is similar as in package \code{reshape2}.
For \code{mlydata}, a \code{zoo} object can be returned.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c('x', 'y', 'z', 'a'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
melt(zc)
x <- matrix(1 : 36, nrow = 3, byrow = TRUE)
md <- mlydata(x, year = 1991 : 1993, month = 1 : 12)
melt(md)
melt(md, ret = 'zoo')
}

