% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{cbind.mlydata}
\alias{cbind.zoocat}
\alias{merge}
\alias{merge.mlydata}
\alias{merge.zoocat}
\title{Combine \code{zoocat} or \code{mlydata} Objects by Columns}
\usage{
## S3 method for class "zoocat"
merge(...)

## S3 method for class "mlydata"
merge(...)

cbind.zoocat(...)

cbind.mlydata(...)
}
\arguments{
\item{...}{\code{zoocat} or \code{mlydata} objects.}
}
\value{
\code{merge.zoocat} will return a \code{zoocat} object. \code{merge.mlydata} will return a \code{zoo} object.
}
\description{
Combine \code{zoocat} or \code{mlydata} objects by columns.
}
\details{
merge.zoocat is similar with merge.zoo, and the column attributes are also
combined. For the combination of cattr, some NA will be filled in if it is
necessary. merge.mlydata will return a \code{zoo} object.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = 'xxx')
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
x2 <- x + 100
colAttr2 <- data.frame(modified = TRUE, month = c(4, 6, 7, 9))
zc2 <- zoocat(x2, order.by = 1991 : 1995, colattr = colAttr2)
zc3 <- merge(zc, zc2)
zc4 <- cbind(zc, zc2)
cattr(zc3)

x <- matrix(1 : 20, nrow = 5)
md1 <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
md2 <- adjust_ym(md1, k = 1)
merge(md2, md1)
cbind(md2, md1)
}

