% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlydata.R
\name{mlydata}
\alias{[.mlydata}
\alias{as.matrix.mlydata}
\alias{coredata.mlydata}
\alias{mlydata}
\alias{mon.mlydata}
\alias{mon<-.mlydata}
\alias{print.mlydata}
\alias{yr.mlydata}
\alias{yr<-.mlydata}
\title{mlydata Class}
\usage{
mlydata(x, year, month = 1:12)

\method{print}{mlydata}(x)

\method{mon}{mlydata}(x)

\method{yr}{mlydata}(x)

\method{mon}{mlydata}(x) <- value

\method{yr}{mlydata}(x) <- value

\method{coredata}{mlydata}(x)

\method{as.matrix}{mlydata}(x)

\method{[}{mlydata}(x, i, j, ...)
}
\arguments{
\item{x}{For \code{mlydata} function, it is a matrix or a vector. If x is
a matrix, each row will be treated as a year. If x is a vector, it will be
treated as a matrix with only one column. For other methods, x is a
\code{mlydata} object.}

\item{year}{A numeric vector representing years.}

\item{month}{A numeric vector in which numbers are used to represent months.
The elements of \code{month} can be smaller than 1 or larger than 12, which
means the months of past years or future years. For example, the months of
last year is (1 : 12) - 12.}
}
\value{
\code{mlydata} returns a mlydata object.\cr Using methods
\code{yr} and \code{mon} to get and set the year and month vector.\cr
\code{window} returns a new mlydata object.\cr \code{coredata} return the
core matrix in the object.\cr \code{as.matrix} return the matrix with dim
names.\cr \code{cbind} will return a \code{zoo} object.
}
\description{
A class designed for monthly data.
}
\details{
\code{mlydata} class is dependent on \code{zoo} class. An attribute
"month" is added for storing the month information.\cr For \code{print}
method, each column name is set as month abbreviation + month number. For
example, the column name corresponding to Feb of current year is "Feb.2".
For a negative month number, "neg" is added to instead of "-". For example,
the column name corresponding to Feb of last year is Feb.neg10. When all
month numbers are between 1 and 12, only the abbreviations of months is
printed.\cr Using \code{yr} and \code{mon} to get and set the year and the
month of the object.\cr Method \code{window} return a \code{mlydata}
object of the given year and month. NULL value of year or month means
all.\cr
}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
coredata(md)
mon(md)
mon(md) <- mon(md) + 1
yr(md)
yr(md) <- yr(md) + 1
md[3, 4]

x <-  matrix(1 : 36, nrow = 3)
md <- mlydata(x, year = 1991 : 1993)
}

