% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlydataList.R
\name{mlydataList}
\alias{mlydataList}
\title{mlydataList Class}
\usage{
mlydataList(...)
}
\arguments{
\item{...}{A series of \code{mlydata} object.
The first element can be a list.
For this case, each elements of the list must be a \code{mlydata} object,
and the list must has
names for each elements, and followed arguments will be neglected.}
}
\value{
A \code{mlydataList} object.
}
\description{
mlydataList class.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
x <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
y <- x + 1
mlydataList(list(x = x, y = y))
mlydataList(x, y)
mlydataList(x)
}

