% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlydataList.R
\name{mlydataList}
\alias{mlydataList}
\title{\code{mlydataList} Class}
\usage{
mlydataList(...)
}
\arguments{
\item{...}{a series of \code{mlydata} object or a list of \code{mlydata} objects.
These can be given as named arguments.
If the first argument is a list, followed arguments will be neglected.
Note that the list must be named.}
}
\value{
A \code{mlydataList} object.
}
\description{
A \code{mlydataList} object is a list, and each element of it is
a \code{mlydata} object.
}
\details{
Many methods for class \code{mlydata} are redefined for
\code{mlydataList} class. When call these methods, each element in
the \code{mlydataList} object is transfered to these methods, and
the results are returned as a \code{mlydataList} object.\cr
These methods include:\cr
\code{\link[=window.mlydata]{window}},
\code{\link[=expand.mlydata]{expand}},
\code{\link[=normalize.zoo]{normalize}},
\code{\link[=cummin]{cummin}},
\code{\link[=cumprod]{cumprod}},
\code{\link[=diff]{diff}},
\code{\link[=head]{head}},
\code{\link[=tail]{tail}},
\code{\link[=lag.mlydata]{lag}},
\code{\link[=na.aggregate]{na.aggregate}},
\code{\link[=na.approx]{na.approx}},
\code{\link[=na.contiguous]{na.contiguous}},
\code{\link[=na.spline]{na.spline}},
\code{\link[=na.trim]{na.trim}},
\code{\link[=rollapply]{rollapply}},
\code{\link[=rollmax]{rollmax}},
\code{\link[=rollmean]{rollmean}},
\code{\link[=rollmedian]{rollmedian}},
\code{\link[=rollsum]{rollsum}},
\code{\link[=scale]{scale}}
}
\examples{
x <- matrix(1 : 20, nrow = 5)
x <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
y <- x + 1
mlydataList(list(md1 = x, md2 = y))
mlydataList(x, y)
mlydataList(x)
}

