% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlydataList.R
\name{mlydataList}
\alias{mlydataList}
\title{\code{mlydataList} Class}
\usage{
mlydataList(...)
}
\arguments{
\item{...}{a series of \code{mlydata} object or a list of \code{mlydata} objects.
These can be given as named arguments.
If the first argument is a list, followed arguments will be neglected.
Note that the list must be named.}
}
\value{
A \code{mlydataList} object.
}
\description{
A \code{mlydataList} object is a list, each element of which is
a \code{mlydata} object.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
x <- mlydata(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
y <- x + 1
mlydataList(list(x = x, y = y))
mlydataList(x, y)
mlydataList(x)
}

