% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\alias{normalize.default}
\alias{normalize.zoo}
\title{Normalize data}
\usage{
normalize(x, ...)

\method{normalize}{default}(x, method = "sd1", base.period = 1:nrow(x), ...)

\method{normalize}{zoo}(x, method = "sd1", base.period = index(x), ...)
}
\arguments{
\item{x}{a vector, matrix, data frame or \code{zoo} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{method}{a character string indicating which method to use. 
Must be 'sd1'(default), 'anomaly' or 'perc'. See details.}

\item{base.period}{a vector indicating the index or range of the base period.
If NULL, base period is the all index range. For matrix, base.period means
the row numbers.}
}
\description{
Normalize data or calculate anomaly or anomaly percentage.
}
\details{
Three methods for normalization can be used.
When method is "anomaly", each column is normalized by \eqn{x - M}, 
where \eqn{M} is the mean value based on the \code{base.period}.
When method is "perc", each column is normalized by \eqn{(x - M) / M}.
When method is "sd1", each column is normalized by \eqn{(x - M) / S}, 
where \eqn{S} is the standard deviation based on the \code{base.period}.
}
\examples{

x <- matrix(1 : 20, nrow = 10)
colnames(x) <- c('a', 'b')
rownames(x) <- 1 : 10
normalize(x, method = 'anomaly')
normalize(x, method = 'perc')
normalize(x, method = 'sd1')

z <- zoo(x, order.by = 1991 : 2010)
normalize(z)

x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c(rep('xxx', 3), 'yyy'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
normalize(zc)

}

