% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_mlydataList.R, R/normalize.R
\name{normalize.mlydataList}
\alias{normalize}
\alias{normalize.default}
\alias{normalize.mlydataList}
\alias{normalize.zoo}
\title{Normalize Data}
\usage{
\method{normalize}{mlydataList}(x, ...)

normalize(x, ...)

\method{normalize}{default}(x, mode = "anomaly", ...)

\method{normalize}{zoo}(x, mode = "anomaly", base.period = NULL, ...)
}
\arguments{
\item{x}{A vector, matrix, data frame or zoo object.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{mode}{\code{perc} means percentage of anomaly. \code{sd1} means
normalized anomaly.}

\item{base.period}{If NULL, base period is the all index range.}
}
\description{
Normalize Data
}
\examples{
x <- matrix(rnorm(36, 10, 4), nrow = 3)
normalize(x)
normalize(x, mode = 'perc')
normalize(x, mode = 'sd1')

md <- mlydata(x, year = 1991 : 1993, month = 1 : 12)
normalize(md)
normalize(md, mode = 'perc')
normalize(md, mode = 'sd1')
normalize(as.zoocat(md), mode = 'sd1')
normalize(as.zoocat(md), mode = 'sd1', base.period = 1991 : 1992)
}

