% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R, R/z_zoomlyList_methods.R
\name{normalize}
\alias{normalize}
\alias{normalize.default}
\alias{normalize.zoo}
\alias{normalize.zoomlyList}
\title{Normalize Data}
\usage{
normalize(x, ...)

\method{normalize}{default}(x, method = "sd1", ...)

\method{normalize}{zoo}(x, method = "sd1", base.period = NULL, ...)

\method{normalize}{zoomlyList}(x, ...)
}
\arguments{
\item{x}{a vector, matrix, data frame or \code{zoo} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{method}{a character string indicating which method to use.
"sd1" (default) means making stardard deviation to be one;
"anomaly" means anomaly (departure);
"perc" means percentage of anomaly.}

\item{base.period}{a vector. If NULL, base period is the all index range.}
}
\description{
Normalize data or calculate anomaly or anomaly percentage.
}
\examples{
x <- matrix(rnorm(36, 10, 4), nrow = 3)
normalize(x)
normalize(x, method = 'perc')
normalize(x, method = 'sd1')

zm <- zoomly(x, year = 1991 : 1993, month = 1 : 12)
normalize(zm)
normalize(zm, method = 'perc')
normalize(zm, method = 'sd1')
normalize(as.zoocat(zm), method = 'sd1')
normalize(as.zoocat(zm), method = 'sd1', base.period = 1991 : 1992)
}

