% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\alias{normalize.default}
\alias{normalize.zoo}
\title{Normalize Data}
\usage{
normalize(x, ...)

\method{normalize}{default}(x, method = "sd1", ...)

\method{normalize}{zoo}(x, method = "sd1", base.period = NULL, ...)
}
\arguments{
\item{x}{a vector, matrix, data frame or \code{zoo} object.}

\item{...}{additional arguments to be passed to or from methods.}

\item{method}{a character string indicating which method to use.
"sd1" (default) means making stardard deviation to be one;
"anomaly" means anomaly (departure);
"perc" means percentage of anomaly.}

\item{base.period}{a vector. If NULL, base period is the all index range.}
}
\description{
Normalize data or calculate anomaly or anomaly percentage.
}
\examples{
x <- matrix(rnorm(36, 10, 4), nrow = 3)
normalize(x)
normalize(x, method = 'perc')
normalize(x, method = 'sd1')
}

