% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\alias{prcomp.zoocat}
\title{Principal components analysis of a zoocat object}
\usage{
\method{prcomp}{zoocat}(x, ...)
}
\arguments{
\item{x}{the zoocat object.}

\item{...}{other argument.}
}
\description{
Principal components analysis of a zoocat object
}
\examples{
x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c(rep('xxx', 3), 'yyy'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
pca <- prcomp(zc)
print(pca$rotation)
print(pca$z)

}

