% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_col.R
\name{reprocess_month}
\alias{reprocess_month}
\title{Reprocess month of zoomly}
\usage{
reprocess_month(x, month)
}
\arguments{
\item{x}{a \code{zoomly} object.}

\item{month}{new setting month vector.}
}
\description{
Reprocess month of zoomly
}
\examples{
mat <- matrix(1:48, ncol = 12)
ctable <- data.frame(month = rep(1 : 12))
zm <- zoomly(mat, order.by = 1991 : 1994, colattr = ctable)
reprocess_month(zm, month = -11:2)
reprocess_month(zm, month = -24:3)
}

