% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollmean_by_month.mlydata.R
\name{rollmean_by_month}
\alias{rollmean_by_month}
\alias{rollmean_by_month.mlydata}
\title{The Rolling Means of a mlydata Object by Month}
\usage{
rollmean_by_month(x, ...)

\method{rollmean_by_month}{mlydata}(x, k, onlyUsePrev = TRUE)
}
\arguments{
\item{x}{A \code{mlydata} object.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{k}{The width of the rolling window.}

\item{onlyUsePrev}{If TRUE, the rolling window is asymmetric, and only}
}
\value{
A mlydata object.\cr
}
\description{
The rolling means of a mlydata object by month.
}
\details{
previous and current data is in the rolling window.
}
\examples{
x <- matrix(1 : 60, nrow = 5)
md <- mlydata(x, year = 1991 : 1995)
rollmean_by_month(md, k = 2)
rollmean_by_month(md, k = 3, onlyUsePrev = FALSE)
rollmean(md, k = 3)
}

