% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spdf2mat.R
\name{spdf2mat}
\alias{spdf2mat}
\title{Translate Data Frame of Spatial Data to Matrix}
\usage{
spdf2mat(x, lat.var = "lat", lon.var = "lon", value.var = "value")
}
\arguments{
\item{x}{A data frame.}

\item{lat.var}{Name of column which store latitude.}

\item{lon.var}{Name of column which store longitude.}

\item{value.var}{Name of column which store values.}
}
\value{
A list contain matrix and bbox.
}
\description{
Translate a data frame of spatial data to a matrix. The data frame must contain
columns of latitude, longitude and values.
}
\examples{
x <- data.frame(lat = rep(1:10, each = 10),
                lon = rep(1:10, 10),
                value = 1:100)
mat <- spdf2mat(x)
}

