% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoocat.R
\name{zoocat}
\alias{zoocat}
\title{"\code{zoocat}" Class}
\usage{
zoocat(x = NULL, order.by = index(x), colattr = NULL,
  index.name = "index", ...)
}
\arguments{
\item{x}{a matrix. If \code{x} is a data frame, it will be converted to a matrix.}

\item{order.by}{an index vector with unique entries by which the observations
in x are ordered.}

\item{colattr}{the column attributes. Must be a data frame with column names.}

\item{index.name}{the name of the index variable.}

\item{...}{other arguments for \code{zoo}.}
}
\value{
A \code{zoocat} object.
}
\description{
A class designed for "\code{zoo}" class with a column attribute (\bold{cattr}) table.
}
\details{
"\code{zoocat}" is a class based on the "\code{zoo}" class, which means
"\code{zoo}" with column(C) attributes(AT). An attribute named "\bold{cattr}" is
added in the object, so there are two important attributes in a "\code{zoocat}"
object: "\code{index}" and "\code{cattr}". "\code{index}" is inherited from 
the "\code{zoo}" class, which keeps the row's index, while "\code{cattr}", 
which is a data.frame, keeps the attribute informations of columns. 
The row number of the "\code{cattr}" must be the same with the number of 
columns of the data matrix, and each row of "\code{cattr}" is correspondent to 
each column of the data matrix. The "\code{cattr}" must have column names, 
and each column of "\code{cattr}" means an attribute field of the columns of the 
core data.\cr
In summary, "\code{zoocat}" class can manage the information of each column more 
conveniently than only using column names.
\cr When printing object, column names will be added automatically.\cr 
As for "\code{zoo}" object, \code{coredata} can be used to extract the core data
matrix or data frame from the object.\cr 
\code{as.matrix} returns a matrix with column and row names. The row names is 
the index of the "\code{zoocat}" object, and the column names is based on 
the "\code{cattr}".
}
\examples{

x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c(rep('xxx', 3), 'yyy'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr)
unclass(zc)
zc[1, 3]
zc[2, ]
zc[, '2_xxx']
coredata(zc)
as.matrix(zc)

x <- matrix(1 : 20, nrow = 5)
colAttr <- data.frame(month = c(2, 3, 5, 6), name = c(rep('xxx', 3), 'yyy'))
zc <- zoocat(x, order.by = 1991 : 1995, colattr = colAttr, frequency = 1)

}

