% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoomly.R
\name{zoomly}
\alias{mon}
\alias{mon.zoomly}
\alias{mon<-}
\alias{mon<-.zoomly}
\alias{yr}
\alias{yr.zoomly}
\alias{yr<-}
\alias{yr<-.zoomly}
\alias{zoomly}
\title{\code{zoomly} Class}
\usage{
zoomly(x, year, month = 1:12, order.by = NULL)

mon(x)

mon(x) <- value

yr(x)

yr(x) <- value

\method{mon}{zoomly}(x)

\method{yr}{zoomly}(x)

\method{mon}{zoomly}(x) <- value

\method{yr}{zoomly}(x) <- value
}
\arguments{
\item{x}{a matrix or a vector for function \code{zoomly}. Otherwise, it is
a \code{zoomly} object.
For \code{zoomly}, if x is a matrix, each row will be treated as a year.
If x is a vector, it will be
treated as a matrix with only one column.}

\item{year}{a numeric vector representing years.}

\item{month}{a numeric vector in which numbers are used to represent months.
The elements of \code{month} can be smaller than 1 or larger than 12, which
means the months of past years or future years. For example, the months of
last year is (1 : 12) - 12.}

\item{order.by}{the same as year. If it is not NULL, argument \code{year} will
be negleted.}

\item{value}{The new value for month and year of the object.}
}
\value{
\code{zoomly} returns a \code{zoomly} object.
}
\description{
A class designed for monthly data.
}
\details{
\code{zoomly} class is dependent on \code{zoo} class. An attribute
"month" is added for storing the month information.\cr For \code{print}
method, each column name is set as month abbreviation + month number. For
example, the column name corresponding to Feb of current year is "Feb.2".
When all month numbers are between 1 and 12, only the abbreviations of months is
printed.\cr Using \code{yr} and \code{mon} to get and set the year and the
month of the object.
}
\examples{
x <- matrix(1 : 20, nrow = 5)
md <- zoomly(x, year = 1991 : 1995, month = c(2, 3, 5, 6))
md <- zoomly(x, order.by = 1991 : 1995, month = c(2, 3, 5, 6))
mon(md)
mon(md) <- mon(md) + 1
yr(md)
yr(md) <- yr(md) + 1

x <-  matrix(1 : 36, nrow = 3)
md <- zoomly(x, year = 1991 : 1993)
}

